Add files via upload
Addition of the index.html and default.py. The index.html is the view layer, and is dependant on the default.py pytjon file. All of the code from map_maker_V2 and MMU_Hack_Crim.Area.py have been aggregated in this folder to create a UI that can do anywhere in the UK.